refactor(sorthelper): moves filter sorting to its own file
